Newish build
Did some things to try and make an object appear runs and seems all the
code is complying but nothing is appearing on screen. Thought it might
be a good boost at least to the project. Made it as a seperate folder
due to it being possibly scrapped if it isn't what we want.